requirements_compat.txt
asgiref
Django<6.0
django-cors-headers
django-filter
djangorestframework
djangorestframework_simplejwt
PyJWT
sqlparse
tzdata
